Cordial.Core.Types

Description

Exports core types for everything from end-user bot config to those needed to write custom modules.

Documentation

data CordialConfig

Basic, primitive config for the bot.

Always use defConfig and alter it as necessary, so that as fields get added here, your bot continues to work without changes.

Constructors

CordialConfig

Fields

botComchar :: Text

botToken :: Text

botModules :: [Module]

defConfig :: CordialConfig

The default config for a Cordial bot.

At a minimum, you must override botToken.

data Callback

The type of a callback that can be registered by modules.

Callbacks are sent every event that the bot becomes aware of, and it is up to the modules to implement the cases they care about in their callbacks *and remember to gracefully disregard the rest*.

Callbacks get fired *before* command handling is executed.

Constructors

Callback (Dis -> Event -> IO ())

data Module

Basic information for a module.

Modules extend the functionality of Cordial and allow it to react to any type of event that is necessary. They get registered in the bot's CordialConfig.

Constructors

Module

Fields

name :: Text

A short name for the module.

description :: Text

A short description of the module.

commands :: [ModuleCommand]

Commands the module exports.

callbacks :: [Callback]

Callbacks that events get fired to.

type Dis = (RestChan, Gateway, [ThreadIdType])

Mostly used internally, just to make it easier to work with the Discord library. Exported for module convenience.

data ModuleCommand

A command that is part of a module.

Constructors

ModuleCommand

Fields

command :: Text

The command name itself.

commandHelp :: Maybe Text

Help/usage text for it.

func :: CordialConfig -> Dis -> Message -> IO ()

The actual implementation of the command.