Cordial.Core.Utility

Description

Useful utility functions that modules can make use of for doing various things. Think of this as a tiny stdlib.

Documentation

args :: Message -> [Text]

Parse out the arguments of a message, dropping the first word (usually the command name with comchar).

sendMsg :: Dis -> Snowflake -> Text -> IO ()

Send a message to a particular channel.

reply :: Dis -> Message -> Text -> IO ()

Reply to a message (takes a message and sends a message to the channel from which it came).

createDM :: Dis -> User -> IO (Either String Snowflake)

Create a DM with a user.