Cordial.Module.Gatekeeper

Description

A module implementing basic authorization. Do not rely on it for now.

Also, it is useless for everyone except YPH/YSU right now.

Documentation

data PendingConfirmation

Someone who has asked for verification but has yet to verify.

Constructors

PendingConfirmation

Fields

pcUserid :: Snowflake

pcKey :: Text

pcValue :: Text

data GatekeeperPredicate

A predicate that the argument to the initial verification command must match.

Constructors

MustEndIn [Text]

processPredicate :: Text -> GatekeeperPredicate -> Bool

Process a given predicate over a given Text.

gatekeeperModule :: [GatekeeperPredicate] -> Module

This module. Hello!

gkPredicates :: [GatekeeperPredicate] -> ModuleCommand

A command that shows the predicates that the argument must match.

confirmid :: [GatekeeperPredicate] -> ModuleCommand

The command to start verification.

greetUser :: Callback

A callback that sends a message to a user when they join.